scripts/synthetic_data/image_processor.py
"""Process and validate StyleGAN2-generated images."""

import os
import numpy as np
import tensorflow as tf
import yaml
import logging
from pathlib import Path
from tqdm import tqdm
import cv2
from PIL import Image

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class SyntheticImageProcessor:
    def __init__(self, config_path):
        """Initialize synthetic image processor.
        
        Args:
            config_path (str): Path to configuration file
        """
        with open(config_path) as f:
            self.config = yaml.safe_load(f)
            
        self.face_cascade = cv2.CascadeClassifier(
            cv2.data.haarcascades + 'haarcascade_frontalface_default.xml'
        )
    
    def process_directory(self, input_dir, output_dir):
        """Process all images in a directory.
        
        Args:
            input_dir (str): Directory containing synthetic images
            output_dir (str): Directory to save processed images
        """
        os.makedirs(output_dir, exist_ok=True)
        
        # Get list of images
        image_files = list(Path(input_dir).glob('*.png'))
        logger.info(f"Found {len(image_files)} images to process")
        
        # Process each image
        valid_images = []
        for image_file in tqdm(image_files):
            if self.process_image(image_file, output_dir):
                valid_images.append(image_file)
                
        logger.info(f"Successfully processed {len(valid_images)} images")
        
        # Save list of valid images
        with open(os.path.join(output_dir, 'valid_images.txt'), 'w') as f:
            for image_file in valid_images:
                f.write(f"{image_file.name}\n")
    
    def process_image(self, image_path, output_dir):
        """Process a single synthetic image.
        
        Args:
            image_path (Path): Path to input image
            output_dir (str): Directory to save processed image
            
        Returns:
            bool: True if processing was successful
        """
        try:
            # Load image
            image = cv2.imread(str(image_path))
            if image is None:
                logger.warning(f"Failed to load image: {image_path}")
                return False
            
            # Validate image quality
            if not self._validate_image_quality(image):
                logger.warning(f"Image failed quality validation: {image_path}")
                return False
            
            # Detect face
            if not self._validate_face_detection(image):
                logger.warning(f"No valid face detected: {image_path}")
                return False
            
            # Save processed image
            output_path = os.path.join(output_dir, image_path.name)
            cv2.imwrite(output_path, image)
            
            return True
            
        except Exception as e:
            logger.error(f"Error processing {image_path}: {str(e)}")
            return False
    
    def _validate_image_quality(self, image):
        """Validate image quality metrics.
        
        Args:
            image (numpy.ndarray): Input image
            
        Returns:
            bool: True if image meets quality standards
        """
        # Check resolution
        height, width = image.shape[:2]
        min_dimension = min(height, width)
        if min_dimension < 64:  # Minimum size threshold
            return False
        
        # Check contrast
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
        contrast = gray.std()
        if contrast < 20:  # Minimum contrast threshold
            return False
        
        # Check for artifacts
        blurriness = cv2.Laplacian(gray, cv2.CV_64F).var()
        if blurriness < 100:  # Minimum sharpness threshold
            return False
            
        return True
    
    def _validate_face_detection(self, image):
        """Validate face detection in image.
        
        Args:
            image (numpy.ndarray): Input image
            
        Returns:
            bool: True if valid face is detected
        """
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
        faces = self.face_cascade.detectMultiScale(
            gray,
            scaleFactor=1.1,
            minNeighbors=5,
            minSize=(30, 30)
        )
        
        # Check if exactly one face is detected
        if len(faces) != 1:
            return False
            
        # Validate face size relative to image
        x, y, w, h = faces[0]
        face_area = w * h
        image_area = image.shape[0] * image.shape[1]
        face_ratio = face_area / image_area
        
        # Face should occupy reasonable portion of image
        return 0.1 <= face_ratio <= 0.9

def main():
    """Main function to run synthetic image processing."""
    import argparse
    
    parser = argparse.ArgumentParser(description="Process synthetic facial expression data")
    parser.add_argument("--config", type=str, default="configs/stylegan_config.yaml",
                      help="Path to configuration file")
    parser.add_argument("--input_dir", type=str, required=True,
                      help="Directory containing synthetic images")
    parser.add_argument("--output_dir", type=str, required=True,
                      help="Directory to save processed images")
    args = parser.parse_args()
    
    # Initialize processor
    processor = SyntheticImageProcessor(args.config)
    
    # Process images
    processor.process_directory(args.input_dir, args.output_dir)
    
if __name__ == "__main__":
    main()