requirements.txt
# Deep Learning Framework
tensorflow-gpu==2.9.1
torch==1.12.1
torchvision==0.13.1

# Computer Vision & Image Processing
opencv-python==4.6.0.66
Pillow==9.2.0
scikit-image==0.19.3
albumentations==1.3.0

# Machine Learning & Data Processing
numpy==1.23.3
pandas==1.4.4
scikit-learn==1.1.2

# StyleGAN2 Requirements
ninja==1.10.2
tensorflow-gan==2.0.0

# Video Processing
pytube==12.1.0
moviepy==1.0.3
pafy==0.5.5
youtube-dl==2021.12.17

# Utilities
tqdm==4.64.1
PyYAML==6.0
matplotlib==3.5.3
seaborn==0.12.0

# Development Tools
jupyter==1.0.0
ipython==8.5.0

# Optional: Quality & Testing
pytest==7.1.3
black==22.8.0
flake8==5.0.4